Move BaseResponse
